requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
schedule>=1.2.0
python-telegram-bot>=20.0
pyupbit>=0.5.1 